+++ src/main/java/org/server/remoteclass/controller/OrderController.java
@@ -3,7 +3,6 @@ package org.server.remoteclass.controller;
 
 import io.swagger.annotations.ApiOperation;
 import org.server.remoteclass.dto.order.RequestOrderDto;
-import org.server.remoteclass.dto.order.ResponseOrderByAdminDto;
 import org.server.remoteclass.dto.order.ResponseOrderDto;
 import org.server.remoteclass.exception.ForbiddenException;
 
@@ -48,23 +47,4 @@ public class OrderController {
     public ResponseEntity<List<ResponseOrderDto>> getMyOrder() throws IdNotExistException {
         return ResponseEntity.status(HttpStatus.OK).body(orderService.getMyOrdersByUserId());
     }
-    @ApiOperation("관리자의 주문 전체 목록 조회")
-    @GetMapping("/admin/list")
-    public ResponseEntity<List<ResponseOrderByAdminDto>> getAllByAdmin() throws IdNotExistException, ForbiddenException {
-        return ResponseEntity.status(HttpStatus.OK).body(orderService.getAllOrdersByAdmin());
-    }
-
-    @ApiOperation("관리자의 사용자별 목록 조회")
-    @GetMapping("/admin/user/{userId}")
-    public ResponseEntity<List<ResponseOrderByAdminDto>> getByUserIdByAdmin(@PathVariable("userId") Long userId) throws IdNotExistException, ForbiddenException {
-        return ResponseEntity.status(HttpStatus.OK).body(orderService.getOrderByUserIdByAdmin(userId));
-    }
-
-    @ApiOperation("관리자의 특정 주문 목록 조회")
-    @GetMapping("/admin/order/{orderId}")
-    public ResponseEntity<ResponseOrderByAdminDto> getByOrderIdByAdmin(@PathVariable("orderId") Long orderId) throws IdNotExistException, ForbiddenException {
-        return ResponseEntity.status(HttpStatus.OK).body(orderService.getOrderByOrderIdByAdmin(orderId));
-    }
-
-
 }
